moving tests/util to testing/
