+++ testing/mocks.py
@@ -1,6 +1,5 @@
-"""
- This is a collection of utility functions for easier, DRY testing.
-"""
+"""This is a collection of utility functions for easier, DRY testing."""
+import io
 from collections import namedtuple
 from contextlib import contextmanager
 from subprocess import CalledProcessError
@@ -31,7 +30,7 @@ def mock_git_calls(subprocess_namespace, cases):
 
         try:
             case = cases[current_case['index']]
-        except IndexError:
+        except IndexError:  # pragma: no cover
             raise AssertionError(
                 '\nExpected: ""\n'
                 'Actual: "{}"'.format(
@@ -40,7 +39,7 @@ def mock_git_calls(subprocess_namespace, cases):
             )
         current_case['index'] += 1
 
-        if command != case.expected_input:
+        if command != case.expected_input:  # pragma: no cover
             # Pretty it up a little, for display
             if not case.expected_input.startswith('git'):
                 case.expected_input = 'git ' + case.expected_input
@@ -120,6 +119,10 @@ def mock_open(data, namespace):
         m().seek = m.side_effect
 
         yield m
+
+
+def mock_file_object(string):
+    return io.StringIO(string)
 
 
 @contextmanager
