+++ tests/core/secrets_collection_test.py
@@ -13,9 +13,9 @@ from detect_secrets.core.secrets_collection import SecretsCollection
 from detect_secrets.plugins.base import BasePlugin
 from detect_secrets.plugins.high_entropy_strings import HexHighEntropyString
 from detect_secrets.plugins.private_key import PrivateKeyDetector
-from tests.util.factories import secrets_collection_factory
-from tests.util.mock_util import mock_log as mock_log_base
-from tests.util.mock_util import mock_open as mock_open_base
+from testing.factories import secrets_collection_factory
+from testing.mocks import mock_log as mock_log_base
+from testing.mocks import mock_open as mock_open_base
 
 
 @pytest.fixture
