+++ tests/plugins/private_key_test.py
@@ -4,7 +4,7 @@ from __future__ import unicode_literals
 import pytest
 
 from detect_secrets.plugins.private_key import PrivateKeyDetector
-from tests.util.file_util import create_file_object_from_string
+from testing.mocks import mock_file_object
 
 
 class TestPrivateKeyDetector(object):
@@ -27,7 +27,7 @@ class TestPrivateKeyDetector(object):
     def test_analyze(self, file_content):
         logic = PrivateKeyDetector()
 
-        f = create_file_object_from_string(file_content)
+        f = mock_file_object(file_content)
         output = logic.analyze(f, 'mock_filename')
         assert len(output) == 1
         for potential_secret in output:
